More scripting cleanup
