+++ src/client/game/scripting/entity.hpp
@@ -1,10 +1,9 @@
 #pragma once
 #include "game/game.hpp"
+#include "script_value.hpp"
 
 namespace scripting
-{
-	class script_value;
-	
+{	
 	class entity final
 	{
 	public:
@@ -20,7 +19,12 @@ namespace scripting
 		entity& operator=(entity&& other) noexcept;
 
 		void set(const std::string& field, const script_value& value) const;
-		script_value get(const std::string& field) const;
+
+		template<typename T = script_value>
+		T get(const std::string& field) const
+		{
+			return this->get<script_value>(field).as<T>();
+		}
 
 		script_value call(const std::string& name, const std::vector<script_value>& arguments) const;
 
@@ -33,4 +37,7 @@ namespace scripting
 		void add() const;
 		void release() const;
 	};
+
+	template<>
+	script_value entity::get(const std::string& field) const;
 }
