+++ src/client/game/scripting/execution.hpp
@@ -9,6 +9,15 @@ namespace scripting
 	script_value call_function(const std::string& name, const entity& entity,
 	                           const std::vector<script_value>& arguments);
 
+	template <typename T = script_value>
+	T call(const std::string& name, const std::vector<script_value>& arguments = {})
+	{
+		return call<script_value>(name, arguments).as<T>();
+	}
+
+	template <>
+	script_value call(const std::string& name, const std::vector<script_value>& arguments);
+
 	void set_entity_field(const entity& entity, const std::string& field, const script_value& value);
 	script_value get_entity_field(const entity& entity, const std::string& field);
 }
