+++ src/client/game/scripting/script_value.cpp
@@ -1,5 +1,6 @@
 #include <std_include.hpp>
 #include "script_value.hpp"
+#include "entity.hpp"
 
 
 namespace scripting
